+++ intelmq/lib/harmonization.py
@@ -378,7 +378,10 @@ class FQDN(GenericType):
     def sanitize(value):
         value = value.strip('.')
         if value:
-            return value.encode('idna').decode().lower()
+            try:
+                return value.encode('idna').decode().lower()
+            except UnicodeError:
+                return
 
     @staticmethod
     def to_ip(value):