+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
 - warnings of bots are catched by the logger (#1074)
 - Bots stop when redis gives the error "OOM command not allowed when used memory > 'maxmemory'.".
 - lib/bot: Fixed exitcodes 0 for graceful shutdowns.
+- lib/harmonization: Handle idna encoding error in FQDN sanitation (#1175).
 
 ### Harmonization
 - Rule for harmonization keys is enforced (#1104)